impl Drop for PythonChild instead of SnippetExtractor
